backend/requirements.txt
djangorestframework==3.19.1
Django==4.3.6
channels==4.1.0
channels_redis==4.1.0
djangorestframework-simplejwt==6.1.0
mysqlclient==2.2.0
python-dotenv==1.0.0
django-cors-headers==4.0.0
Pillow==10.0.0
asgiref==3.8.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
